requirements.txt
numpy>=1.23
pandas>=1.5
Pillow>=9.0
requests>=2.28
PyYAML>=6.0
torch>=2.0
torchvision>=0.15
transformers>=4.30
tqdm>=4.64
rapidfuzz>=3.0.0
rdflib>=6.0
nltk>=3.7
wget>=3.2
pytest>=7.0
Jinja2>=3.0
